[TAN-8601] Move LockedNote to the shared content builder widgets
