+++ front/app/components/ProjectPageBuilder/Widgets/ProjectBanner/index.tsx
@@ -16,6 +16,7 @@ import useLocalize from 'hooks/useLocalize';
 import { IMAGE_UPLOADING_EVENT } from 'components/admin/ContentBuilder/constants';
 import useWidgetProjectId from 'components/admin/ContentBuilder/useWidgetProjectId';
 import imageMessages from 'components/admin/ContentBuilder/Widgets/ImageMultiloc/messages';
+import LockedNote from 'components/admin/ContentBuilder/Widgets/LockedNote';
 import ImageCropperContainer from 'components/admin/ImageCropper/Container';
 import {
   HeaderImage,
@@ -30,7 +31,6 @@ import eventEmitter from 'utils/eventEmitter';
 import { convertUrlToUploadFile } from 'utils/fileUtils';
 import { usePermission } from 'utils/permissions';
 
-import LockedNote from '../LockedNote';
 import messages from '../messages';
 
 import EmptyBanner from './EmptyBanner';
