+++ front/app/components/admin/ContentBuilder/Widgets/LockedNote.tsx
